SimString

Documentation for SimString.

A native Julia implementation of the CPMerge algorithm, which is designed for approximate string matching. This package is be particulary useful for natural language processing tasks which demand the retrieval of strings/texts from a very large corpora (big amounts of texts). Currently, this package supports both Character and Word based N-grams feature generations and there are plans to open the package up for custom user defined feature generation methods.

Features

[X] Fast algorithm for string matching

[X] 100% exact retrieval

[X] Support for unicodes

[ ] Custom user defined feature generation methods

[ ] Mecab-based tokenizer support

Suported String Similarity Measures

[X] Dice coefficient

[X] Jaccard coefficient

[X] Cosine coefficient

[X] Overlap coefficient

Installation

You can grab the latest stable version of this package from Julia registries by simply running;

NB: Don't forget to invoke Julia's package manager with ]

pkg> add SimString

The few (and selected) brave ones can simply grab the current experimental features by simply adding the master branch to your development environment after invoking the package manager with ]:

pkg> add SimString#master

You are good to go with bleeding edge features and breakages!

To revert to a stable version, you can simply run:

pkg> free SimString

Usage

using SimString

# Inilisate database and some strings
db = DictDB(CharacterNGrams(2, " "));
push!(db, "foo");
push!(db, "bar");
push!(db, "fooo");

# Convinient approach is to use an array of strings for multiple entries: `append!(db, ["foo", "bar", "fooo"]);`

# Retrieve the closest match(es)
res = search(Dice(), db, "foo"; α=0.8, ranked=true)
# 2-element Vector{Tuple{String, Float64}}:
#  ("foo", 1.0)
#  ("fooo", 0.8888888888888888)

TODO: Benchmarks

Release History

0.1.0 Initial release.

SimString.AbstractSimStringDB

SimString.AbstractSimilarityMeasure

SimString.CharacterNGrams

SimString.Cosine

SimString.Dice

SimString.DictDB

SimString.DictDB

SimString.DictDB

SimString.FeatureExtractor

SimString.Jaccard

SimString.Overlap

SimString.WordNGrams

Base.append!

Base.push!

SimString.cummulative_ngram_count

SimString.extract_features

SimString.extract_features

SimString.init_ngrams

SimString.init_ngrams

SimString.lookup_feature_set_by_size_feature

SimString.maximum_feature_size

SimString.maximum_feature_size

SimString.maximum_feature_size

SimString.maximum_feature_size

SimString.minimum_feature_size

SimString.minimum_feature_size

SimString.minimum_feature_size

SimString.minimum_feature_size

SimString.minimum_overlap

SimString.minimum_overlap

SimString.minimum_overlap

SimString.minimum_overlap

SimString.n_grams

SimString.n_grams

SimString.overlap_join

SimString.pad_string

SimString.pad_string

SimString.rank_search_results

SimString.retrieve_existing_feature_by_size

SimString.search

SimString.search!

SimString.similarity_score

SimString.similarity_score

SimString.similarity_score

SimString.similarity_score

SimString.AbstractSimStringDB — Type

Base type for all custom db collections.

SimString.AbstractSimilarityMeasure — Type

Abstract base type for all string similarity measures.

SimString.CharacterNGrams — Type

Feature extraction on character-level ngrams

SimString.Cosine — Type

Cosine Similarity Measure.

SimString.Dice — Type

Dice Similarity Measure.

SimString.DictDB — Type

Custom DB collection for storing SimString data using base Dictionary Dict

SimString.DictDB — Method

DictDB(x::CharacterNGrams)

Initialize a dict DB with additional containers and Metadata for CharacterNGrams

Arguments

x: CharacterNGrams object

Example

db = DictDB(CharacterNGrams(2, " "))

Returns

DictDB: A DictDB object with additional containers and Metadata for CharacterNGrams

SimString.DictDB — Method

DictDB(x::WordNGrams)

Initialize a dict DB with additional containers and Metadata for WordNGrams

Arguments

x: WordNGrams object

Example

db = DictDB(WordNGrams(2, " ", " "))

Returns

DictDB: A DictDB object with additional containers and Metadata for WordNGrams

SimString.FeatureExtractor — Type

Abstract type for feature extraction structs

SimString.Jaccard — Type

Jaccard Similarity Measure.

SimString.Overlap — Type

Overlap Similarity Measure.

SimString.WordNGrams — Type

Feature extraction based on word-level ngrams

Base.append! — Method

Add bulk items to a new or existing collection of strings using the custom AbstractSimStringDB type.

Base.push! — Method

Add a new item to a new or existing collection of strings using the custom AbstractSimStringDB type.

SimString.cummulative_ngram_count — Method

Internal function to count and pad generated character-level ngrams (including duplicates)

SimString.extract_features — Method

Internal function to generate character-level ngrams features from an AbstractString

SimString.extract_features — Method

Internal function to generate word-level ngrams features from an AbstractString

SimString.init_ngrams — Method

Internal function to generate intial uncounted ngrams on a character level

SimString.init_ngrams — Method

Internal function to generate intial uncounted ngrams on a word level

SimString.lookup_feature_set_by_size_feature — Method

Internal function to lookup feature sets by size and feature

SimString.maximum_feature_size — Method

Calculate maximum feature size for Cosine similarity measure.

SimString.maximum_feature_size — Method

Calculate maximum feature size for Dice similarity measure.

SimString.maximum_feature_size — Method

Calculate maximum feature size for Jaccard similarity measure.

SimString.maximum_feature_size — Method

Calculate maximum feature size for Overlap similarity measure.

SimString.minimum_feature_size — Method

Calculate minimum feature size for Cosine similarity measure.

SimString.minimum_feature_size — Method

Calculate minimum feature size for Dice similarity measure.

SimString.minimum_feature_size — Method

Calculate minimum feature size for Jaccard similarity measure.

SimString.minimum_feature_size — Method

Calculate minimum feature size for Overlap similarity measure.

SimString.minimum_overlap — Method

Calculate the minimum overlap (τ) for a query size, candidate size, and α using Cosine similarity measure.

SimString.minimum_overlap — Method

Calculate the minimum overlap (τ) for a query size, candidate size, and α using Dice similarity measure.

SimString.minimum_overlap — Method

Calculate the minimum overlap (τ) for a query size, candidate size, and α using Jaccard similarity measure.

SimString.minimum_overlap — Method

Calculate the minimum overlap (τ) for a query size, candidate size, and α using Overlap similarity measure.

SimString.n_grams — Method

Internal function to create character-level ngrams features from an AbstractString

SimString.n_grams — Method

Internal function to create word-level ngrams from an AbstractVector

SimString.overlap_join — Method

Internal function which performs the overlap join

SimString.pad_string — Method

Internal function to pad AbstractString types with specified padder

SimString.pad_string — Method

Internal function to pad AbstractVector types with specified padder

SimString.rank_search_results — Method

Internal function which ranks the results of a search using the specified similarity measure.

SimString.retrieve_existing_feature_by_size — Method

Internal function for retrieving existing features by size

SimString.search! — Method

Search for strings in custom DictDB string collection using the SimString algorithm and a similarity measure.

SimString.search — Method

search(measure::AbstractSimilarityMeasure, db_collection::AbstractSimStringDB, query::AbstractString;
    α=0.7, ranked=true)

Search for strings in a string collection using the SimString algorithm and a similarity measure.

Arguments:

measure::AbstractSimilarityMeasure - The similarity measure to use.

db_collection::AbstractSimStringDB - The database collection to search.

query::AbstractString - The query string to search for.

α::float - The α parameter for the SimString algorithm.

ranked::Boolean - Whether to return the results in ranked order.

Example

db = DictDB(CharacterNGrams(2, " "));
append!(db, ["foo", "bar", "fooo"]);

search(Dice(), db, "foo"; α=0.8, ranked=true)
# 2-element Vector{Tuple{String, Float64}}:
#  ("foo", 1.0)
#  ("fooo", 0.8888888888888888)

Returns

A Vector of results, where each element is a Tuple of the form (string, similarity measure score).

SimString.similarity_score — Method

Calculate similarity score between X and Y using Cosine similarity measure.

SimString.similarity_score — Method

Calculate similarity score between X and Y using Dice similarity measure.

SimString.similarity_score — Method

Calculate similarity score between X and Y using Jaccard similarity measure.

SimString.similarity_score — Method

Calculate similarity score between X and Y using Overlap similarity measure.